+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Additions
 - Added UiAnimation system
 - Added AddCustomStyle and ApplyCustomStyle to UiStyle to allow for easy custom styling of elements
 - Added UiControls.PressElement
+- Added TextField.EnterReceiver
 
 Improvements
 - Increased Element area calculation recursion limit to 64